fix: fix error handling.

+++ src/main/java/com/bulish/melnikov/converter/service/ConvertRequestServiceImpl.java
@@ -4,13 +4,16 @@ import com.bulish.melnikov.converter.model.ConvertRequest;
 import com.bulish.melnikov.converter.model.State;
 import com.bulish.melnikov.converter.repository.ConverterRequestRepository;
 import lombok.AllArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
+import java.io.IOException;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
 
 @Service
+@Slf4j
 @AllArgsConstructor
 public class ConvertRequestServiceImpl implements ConvertRequestService {
 
@@ -18,18 +21,18 @@ public class ConvertRequestServiceImpl implements ConvertRequestService {
     private final FileService fileService;
 
     @Override
-    public ConvertRequest save(ConvertRequest convertRequest){
+    public ConvertRequest save(ConvertRequest convertRequest) {
         return requestRepo.save(convertRequest);
     }
 
     @Override
-    public ConvertRequest get(String convertRequestId){
+    public ConvertRequest get(String convertRequestId) {
         return requestRepo.findById(convertRequestId).orElseThrow(()
                 -> new RuntimeException("File not found"));
     }
 
     @Override
-    public ConvertRequest update(ConvertRequest convertRequest){
+    public ConvertRequest update(ConvertRequest convertRequest) {
         ConvertRequest convertRequestFromRedis = get(convertRequest.getId());
 
         convertRequestFromRedis.setState(convertRequest.getState());
@@ -41,11 +44,10 @@ public class ConvertRequestServiceImpl implements ConvertRequestService {
     }
 
     @Override
-    public void delete(String convertRequestId){
+    public void delete(String convertRequestId) {
         requestRepo.deleteById(convertRequestId);
     }
 
-    @Override
     public List<ConvertRequest> getOldConvertRequests() {
         LocalDateTime dateTime = LocalDateTime.now().minusHours(2);
         List<ConvertRequest> convertRequests = new ArrayList<>();
@@ -54,26 +56,35 @@ public class ConvertRequestServiceImpl implements ConvertRequestService {
 
         for (ConvertRequest request : requests) {
 
-            if (request.getState() == State.CONVERTED ||
-                request.getState() == State.IN_ERROR &&
-                request.getConversionDate().isBefore(dateTime)) {
+            if ((request.getState() == State.CONVERTED ||
+                    request.getState() == State.IN_ERROR) &&
+                    request.getConversionDate().isBefore(dateTime)) {
 
                 convertRequests.add(request);
             }
         }
 
-     return convertRequests;
+        return convertRequests;
     }
 
-    @Override
     public void deleteOldConvertRequestWithFiles() {
         List<ConvertRequest> convertRequests = getOldConvertRequests();
+        boolean isDeleted = false;
 
-        for (ConvertRequest request : convertRequests) {
-            fileService.deleteFile(request.getFilePath());
-            if (request.getConvertedFilePath() != null)
-                fileService.deleteFile(request.getConvertedFilePath());
+        try {
+            for (ConvertRequest request : convertRequests) {
+
+                if (request.getConvertedFilePath() != null) {
+                    isDeleted = fileService.deleteFile(request.getFilePath(), request.getConvertedFilePath());
+                } else {
+                    isDeleted = fileService.deleteFile(request.getFilePath());
+                }
+                if (!isDeleted) convertRequests.remove(request);
+            }
+        } catch (IOException e) {
+            log.error("Error while removing the file from local system");
         }
-        requestRepo.deleteAll(convertRequests);
+
+        if (!convertRequests.isEmpty()) requestRepo.deleteAll(convertRequests);
     }
 }
